chore: working sentry instrumentation
removed the analytics package since we are not deploying on vercel
moved the client building for queries to a server action
add conditional instrumentation to production
